test(board): split USB VID/PID tests into tests_usb_vid.rs (LOC gate)
PR #413's CI hit the 1000-LOC gate on board/tests.rs (1005 lines).
Move the entire "USB VID/PID defines" test block (5 tests) into a
sibling sub-module file so the main tests.rs drops to 930 LOC.

No behavior change — same 5 tests, same assertions, just split into a
dedicated file. 72/72 board tests pass.
