+++ crates/fbuild-config/src/board/tests.rs
@@ -700,81 +700,6 @@ fn test_esp32c3_no_psram() {
     );
 }
 
-// --- USB VID/PID defines ---
-
-#[test]
-fn test_get_defines_usb_vid_pid() {
-    let mut config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
-    config.vid = Some("0x2341".to_string());
-    config.pid = Some("0x8036".to_string());
-    let defines = config.get_defines();
-    assert_eq!(defines.get("USB_VID"), Some(&"0x2341".to_string()));
-    assert_eq!(defines.get("USB_PID"), Some(&"0x8036".to_string()));
-}
-
-#[test]
-fn test_get_defines_no_usb_when_absent() {
-    let config = BoardConfig::from_board_id("esp32dev", &HashMap::new()).unwrap();
-    assert!(config.vid.is_none());
-    let defines = config.get_defines();
-    assert!(!defines.contains_key("USB_VID"));
-    assert!(!defines.contains_key("USB_PID"));
-}
-
-#[test]
-fn test_leonardo_board_has_vid_pid() {
-    let config = BoardConfig::from_board_id("leonardo", &HashMap::new()).unwrap();
-    assert_eq!(config.vid, Some("0x2341".to_string()));
-    assert_eq!(config.pid, Some("0x8036".to_string()));
-    let defines = config.get_defines();
-    assert_eq!(defines.get("USB_VID"), Some(&"0x2341".to_string()));
-    assert_eq!(defines.get("USB_PID"), Some(&"0x8036".to_string()));
-}
-
-/// FastLED/fbuild#405: on ESP32-S2/S3 the Arduino framework's
-/// `variants/<board>/pins_arduino.h` already defines USB_VID/USB_PID, so
-/// injecting them again via `-D` produces a "USB_VID redefined" warning at
-/// every TU that includes pins_arduino.h (149-156 per build observed in CI).
-/// The fix: get_defines() must NOT emit USB_VID/USB_PID for ESP32-S2/S3
-/// boards even when the board JSON carries vid/pid fields.
-#[test]
-fn test_esp32s3_board_skips_usb_vid_pid_injection() {
-    let config = BoardConfig::from_board_id("adafruit_feather_esp32s3", &HashMap::new()).unwrap();
-    // The board JSON has vid/pid set (Adafruit Feather ESP32-S3 = 0x239A/0x811B).
-    assert!(
-        config.vid.is_some(),
-        "esp32s3 board should have vid in JSON"
-    );
-    assert!(
-        config.pid.is_some(),
-        "esp32s3 board should have pid in JSON"
-    );
-    // But get_defines() must NOT emit them because the framework variant
-    // header already defines them.
-    let defines = config.get_defines();
-    assert!(
-        !defines.contains_key("USB_VID"),
-        "esp32s3 must not inject USB_VID — framework defines it"
-    );
-    assert!(
-        !defines.contains_key("USB_PID"),
-        "esp32s3 must not inject USB_PID — framework defines it"
-    );
-}
-
-/// User override path: when a user sets `-DUSB_VID=...` via build_flags, the
-/// extra_flags processing should still propagate it. The skip rule only
-/// applies to the unconditional injection from the JSON `vid`/`pid` fields.
-#[test]
-fn test_esp32s3_board_extra_flag_usb_vid_override_wins() {
-    let mut config =
-        BoardConfig::from_board_id("adafruit_feather_esp32s3", &HashMap::new()).unwrap();
-    // Simulate user override via build_flags / board.extra_flags.
-    config.extra_flags = Some("-DUSB_VID=0xDEAD -DUSB_PID=0xBEEF".to_string());
-    let defines = config.get_defines();
-    assert_eq!(defines.get("USB_VID"), Some(&"0xDEAD".to_string()));
-    assert_eq!(defines.get("USB_PID"), Some(&"0xBEEF".to_string()));
-}
 
 #[test]
 fn test_attiny1604_board_config() {
